Changed version system and moved API files
moved api file directory for modding convenience.
update gradle from 2.0 to 4.4.1
