+++ api/api/hbm/energy/IBatteryItem.java
