+++ api/api/hbm/entity/IRadarDetectable.java
