+++ api/api/hbm/entity/IRadiationImmune.java
