+++ build.properties
@@ -1,0 +1,6 @@
+minecraft.version=1.7.10
+forge.version=10.13.4.1614
+hbm.version = 1.0.27-3802B
+nei.version=1.0.5.111
+codechickenlib.version=1.1.3.140
+codechickencore.version=1.0.7.47
