--- src/main/java/cofh/api/CoFHAPIProps.java
@@ -1,11 +1,0 @@
-package cofh.api;
-
-public class CoFHAPIProps {
-
-	private CoFHAPIProps() {
-
-	}
-
-	public static final String VERSION = "1.7.10R1.0.2";
-
-}
