--- src/main/java/cofh/api/energy/IEnergyReceiver.java
@@ -1,38 +1,0 @@
-package cofh.api.energy;
-
-import net.minecraftforge.common.util.ForgeDirection;
-
-/**
- * Implement this interface on Tile Entities which should receive energy, generally storing it in one or more internal {@link IEnergyStorage} objects.
- * <p>
- * A reference implementation is provided {@link TileEnergyHandler}.
- *
- * @author King Lemming
- *
- */
-public interface IEnergyReceiver extends IEnergyConnection {
-
-	/**
-	 * Add energy to an IEnergyReceiver, internal distribution is left entirely to the IEnergyReceiver.
-	 *
-	 * @param from
-	 *            Orientation the energy is received from.
-	 * @param maxReceive
-	 *            Maximum amount of energy to receive.
-	 * @param simulate
-	 *            If TRUE, the charge will only be simulated.
-	 * @return Amount of energy that was (or would have been, if simulated) received.
-	 */
-	int receiveEnergy(ForgeDirection from, int maxReceive, boolean simulate);
-
-	/**
-	 * Returns the amount of energy currently stored.
-	 */
-	int getEnergyStored(ForgeDirection from);
-
-	/**
-	 * Returns the maximum amount of energy that can be stored.
-	 */
-	int getMaxEnergyStored(ForgeDirection from);
-
-}
