--- src/main/java/cofh/api/energy/package-info.java
@@ -1,9 +1,0 @@
-/**
- * (C) 2014 Team CoFH / CoFH / Cult of the Full Hub
- * http://www.teamcofh.com
- */
-@API(apiVersion = CoFHAPIProps.VERSION, owner = "CoFHAPI", provides = "CoFHAPI|energy")
-package cofh.api.energy;
-
-import cofh.api.CoFHAPIProps;
-import cpw.mods.fml.common.API;
