+++ src/main/resources/mcmod.info
@@ -3,8 +3,8 @@
   "modid": "hbm",
   "name": "Hbm's Nuclear Tech",
   "description": "A mod that adds weapons, nuclear themed stuff and machines",
-  "version":"1.0.27-3808",
-  "mcversion": "1.7.10",
+  "version":"${version}",
+  "mcversion": "${mcversion}",
   "url": "",
   "updateUrl": "",
   "authorList": ["HbMinecraft"],
